Run from Read-Only Location - fix dev mode
